:bug: [bug] Fix bug in pipeline publish

+++ dataci/run/save.py
@@ -25,21 +25,22 @@ def save(run: 'Run' = ...):
         #####################################################################
         # Step 1: Recover pipeline feat cached file (.dvc) from .lock
         #####################################################################
-        with open('dvc.lock', 'r') as f:
-            run_cache_lock = yaml.safe_load(f)
-        for k, v in run_cache_lock['stages'].items():
-            for out in v['outs']:
-                logger.info(f'Recover dvc file {out["path"]}.dvc')
-                with open(out['path'] + '.dvc', 'w') as f:
-                    yaml.safe_dump({
-                        'outs': [
-                            {
-                                'md5': out['md5'],
-                                'path': os.path.splitext(out['path'])[0],
-                                'size': out['size']
-                            }
-                        ]
-                    }, f)
+        if os.path.exists('dvc.lock'):
+            with open('dvc.lock', 'r') as f:
+                run_cache_lock = yaml.safe_load(f)
+            for k, v in run_cache_lock['stages'].items():
+                for out in v['outs']:
+                    logger.info(f'Recover dvc file {out["path"]}.dvc')
+                    with open(out['path'] + '.dvc', 'w') as f:
+                        yaml.safe_dump({
+                            'outs': [
+                                {
+                                    'md5': out['md5'],
+                                    'path': os.path.splitext(out['path'])[0],
+                                    'size': out['size']
+                                }
+                            ]
+                        }, f)
 
         #####################################################################
         # Step 2: Publish pipeline output feat
